SAE: Reduce queue wait time for pending Authentication frames
The queue_len * 50 ms wait time was too large with the retransmission
timeouts used in the mesh case for SAE. The maximum wait of 750 ms was
enough to prevent successful completion of authentication after having
hit the maximum queue length. While the previous commit is enough to
allow this to complete successfully in couple of retries, it looks like
a smaller wait time should be used here even if it means potentially
using more CPU.

Drop the processing wait time to queue_len * 10 ms so that the maximum
wait time is 150 ms if the queue is full.

Signed-off-by: Jouni Malinen <jouni@codeaurora.org>

CRs-Fixed: 2414695
Git-repo : git://w1.fi/hostap.git
Git-commit: ff5f54e15941b6809cadd7aeec442f0a16287f7b
Change-Id: I4d31fc6eb3c18c692e542809a7fa30aa013a6b84