Store own MAC address (SPA) in supplicant PMKSA cache entries
This is needed to be able to determine whether a PMKSA cache entry is
valid when using changing MAC addresses. This could also be used to
implement a mechanism to restore a previously used MAC address instead
of a new random MAC address.

Change-Id: I2ce3ac45e74c14de2991fa4423e4c62179e23a4d
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: https://git.w1.fi/hostap.git
Git-commit: 9f04a9c8dd7a023d95ab929648a47c6343c844b8
CRs-Fixed: 4510422