PASN: Add support to reject PASN auth 1 based on user input
When the PASN implementation is used as a library for Wi-Fi Aware
applications there is need to reject the PASN auth 1 frame based on the
user input. Add support to send PASN auth 2 with failure status for
auth 1 frame received to address this need.

Change-Id: I0a934a6bec45fafe00d0af5d3e26e4aa9185136a
Signed-off-by: Jouni Malinen <quic_jouni@quicinc.com>
Git-repo: https://git.w1.fi/hostap.git
Git-commit: 8f21cdf9d765f2d1462fef8f96dcf2e878f938dc
CRs-Fixed: 4510402